Guzzle4 support
