+++ .travis.yml
@@ -4,7 +4,6 @@ php:
   - 5.6
   - 5.5
   - 5.4
-  - 5.3
   - hhvm
 
 install: composer install
