+++ composer.json
@@ -1,13 +1,17 @@
 {
-    "name": "99designs/http-signatures-guzzle",
-    "description": "Sign and verify HTTP messages with Guzzle",
-    "homepage": "https://github.com/99designs/http-signatures-guzzle",
-    "keywords": ["http", "https", "signing", "signed", "signature", "hmac", "guzzle"],
+    "name": "99designs/http-signatures-guzzlehttp",
+    "description": "Sign and verify HTTP messages with Guzzle 4",
+    "homepage": "https://github.com/99designs/http-signatures-guzzlehttp",
+    "keywords": ["http", "https", "signing", "signed", "signature", "hmac", "guzzle 4"],
     "license": "MIT",
     "authors": [
         {
             "name": "Adrian Palmer",
             "email": "adrian.palmer@99designs.com"
+        },
+        {
+            "name": "Ruben de Vries",
+            "email": "ruben@blocktrail.com"
         }
     ],
     "autoload": {
@@ -16,9 +20,9 @@
         }
     },
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=5.4.0",
         "99designs/http-signatures": "~1.1",
-        "guzzle/guzzle": "~3.9"
+        "guzzlehttp/guzzle": "~4.2"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.1"
