+++ src/HttpSignatures/GuzzleHttp/MessageHeaders.php
@@ -1,12 +1,24 @@
 <?php
 
-namespace HttpSignatures\Guzzle;
+namespace HttpSignatures\GuzzleHttp;
 
+use GuzzleHttp\Message\MessageInterface;
+
+/**
+ * Class MessageHeaders
+ *
+ * wrapper around the Guzzle Request instance to have a consistent API for the HttpSignatures classes to consume
+ *
+ * @package HttpSignatures\Guzzle
+ */
 class MessageHeaders
 {
+    /**
+     * @var MessageInterface
+     */
     private $request;
 
-    public function __construct($request)
+    public function __construct(MessageInterface $request)
     {
         $this->request = $request;
     }
